A simple plugin to embed PDFs w/o JS
We use an <object> tag, which is simple, standards-compliant,
and doesn't require JavaScript. The fallback for browsers
without support for embedding PDFs is a link to the PDF file,
and the title, if available.

To embed PDFs which were uploaded to the same Wordpress site
you want to embed them in, you can simply put one of the
following links on a line by itself:
  - URL of the PDF file itself
  - the post short URL
  - the attachment short URL
  - the attachment pretty URL

For PDFs which are on another website, or if you want to use
non-default parameters (width, height, title), use the [pdf]
short tag: [pdf]https://example.com/file.pdf[/pdf]

The plugin also makes the [embed] short tag work for PDF files.

git-svn-id: https://plugins.svn.wordpress.org/vanilla-pdf-embed/trunk@816237 b8457f37-d9ea-0310-8a92-e5e31aec5664